+++ prisma/migrations/20260530041056_auth/migration.sql
@@ -1,0 +1,31 @@
+-- CreateEnum
+CREATE TYPE "OtpProposito" AS ENUM ('verificacion', 'reset_contrasena');
+
+-- CreateTable
+CREATE TABLE "Sesion" (
+    "id" TEXT NOT NULL,
+    "usuarioId" TEXT NOT NULL,
+    "expiraEn" TIMESTAMP(3) NOT NULL,
+    "creadaEn" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+
+    CONSTRAINT "Sesion_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateTable
+CREATE TABLE "CodigoOtp" (
+    "id" TEXT NOT NULL,
+    "usuarioId" TEXT NOT NULL,
+    "hashCodigo" TEXT NOT NULL,
+    "proposito" "OtpProposito" NOT NULL,
+    "expiraEn" TIMESTAMP(3) NOT NULL,
+    "usado" BOOLEAN NOT NULL DEFAULT false,
+    "creadoEn" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+
+    CONSTRAINT "CodigoOtp_pkey" PRIMARY KEY ("id")
+);
+
+-- AddForeignKey
+ALTER TABLE "Sesion" ADD CONSTRAINT "Sesion_usuarioId_fkey" FOREIGN KEY ("usuarioId") REFERENCES "Usuario"("id") ON DELETE CASCADE ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "CodigoOtp" ADD CONSTRAINT "CodigoOtp_usuarioId_fkey" FOREIGN KEY ("usuarioId") REFERENCES "Usuario"("id") ON DELETE CASCADE ON UPDATE CASCADE;